Do some delinting and splitting out of methods/classes/modules.
